[NFC] Move encoding logic from LinalgExt to Codegen/Common
