+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -82,6 +82,7 @@ iree_compiler_cc_library(
         "DecomposeLinalgGeneric.cpp",
         "DecomposePackUnPackOps.cpp",
         "EmulateNarrowType.cpp",
+        "EncodingUtils.cpp",
         "EraseDeadAllocAndStores.cpp",
         "EraseHALDescriptorTypeFromMemRef.cpp",
         "ExtractAddressComputation.cpp",
@@ -121,7 +122,7 @@ iree_compiler_cc_library(
     ],
     hdrs = [
         "BufferizationAnalysis.h",
-        "EncodingInfo.h",
+        "EncodingUtils.h",
         "ExtractAddressComputation.h",
         "Passes.h",
         "TileSizeSelection.h",
