+++ compiler/src/iree/compiler/Codegen/Common/CPU/PassDetail.h
@@ -7,6 +7,7 @@
 #ifndef IREE_COMPILER_CODEGEN_LLVMCPU_PASS_DETAIL_H_
 #define IREE_COMPILER_CODEGEN_LLVMCPU_PASS_DETAIL_H_
 
+#include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
